chore: organise resources
