--- apps/picsa-tools/resources-tool/src/app/data/collections.ts
@@ -1,17 +1,0 @@
-import { IResourceCollection } from '../models';
-
-const collections: { [id: string]: IResourceCollection } = {
-  picsaTrainingVideo: {
-    _created: '2019-09-25T10:00:04.000Z',
-    _modified: '2019-09-27T11:00:01.000Z',
-    _key: 'picsaTrainingVideo',
-    priority: 10,
-    type: 'collection',
-    title: 'PICSA Videos',
-    description: 'Training videos to support PICSA',
-    image: 'assets/resources/covers/ram-refresher.jpg',
-    imageFit: 'cover',
-    childResources: ['m6I8TfROsyr5Wp73BcAd'],
-  },
-};
-export default collections;
