--- apps/picsa-tools/resources-tool/src/app/data/deprecated.ts
@@ -1,34 +1,0 @@
-/**
- * These resources are hard-coded into the app, and therefore should be made available in the
- * assets folder. Note, the order is specified via _created
- * To make changes, the _modified field should also be updated
- */
-
-// {
-//   _key: '9Pkro1VYBUlwuNg5oHok',
-//   _created: '2019-09-25T10:00:03.000Z',
-//   _modified: '2019-09-25T11:00:01.000Z',
-//   title: 'Crop Information',
-//   subtitle: 'Chileka',
-//   filename: 'crop-info-sheet-chileka.pdf',
-//   type: 'pdf',
-//   image: 'assets/resources/crop-info-sheet-chileka-cover.png',
-//   group: 'Documents',
-//   viewableBy: ['wfp-2017'],
-//   url:
-//     'https://firebasestorage.googleapis.com/v0/b/extension-toolkit.appspot.com/o/Resources%2Fcrop-info-sheet-chileka.pdf?alt=media&token=cb8a6243-1d37-43f6-a97c-a0a7bc0f11f2'
-// },
-// {
-//   _key: 'Yx8927IVTGyM1C4njOIv',
-//   _created: '2019-07-20T16:57:13.604Z',
-//   _modified: '2019-07-20T16:57:13.604Z',
-//   _deleted: true,
-//   title: 'Seasonal Forecast',
-//   subtitle: 'Mangochi 2017-2018',
-//   filename: 'seasonal-forecast-mangochi-2017-2018.pdf',
-//   type: 'pdf',
-//   image: 'assets/resources/seasonal-forecast-mangochi-2017-2018.jpg',
-//   group: 'Documents',
-//   url:
-//     'https://firebasestorage.googleapis.com/v0/b/extension-toolkit.appspot.com/o/Resources%2Fseasonal-forecast-mangochi-2017-2018.pdf?alt=media&token=c60180cd-8dcd-4d98-aa7e-48f37dc83849'
-// }
