--- apps/picsa-tools/resources-tool/src/app/data/files.ts
@@ -1,34 +1,0 @@
-import { IResourceFile } from '../models';
-
-const files: Record<string, IResourceFile> = {
-  '6VjjF3yZJejFigwDIPr8': {
-    _key: '6VjjF3yZJejFigwDIPr8',
-    _created: '2019-09-25T10:00:04.000Z',
-    _modified: '2019-09-27T11:00:01.000Z',
-    title: 'Potential Training Schedule',
-    mimetype: 'application/pdf',
-    description: '',
-    subtitle: '',
-    filename: 'potential-PICSA-training-schedule.pdf',
-    type: 'file',
-    image: 'assets/resources/covers/potential-PICSA-training-schedule.png',
-    url: 'https://firebasestorage.googleapis.com/v0/b/extension-toolkit.appspot.com/o/Resources%2Fpotential-PICSA-training-schedule.pdf?alt=media&token=618737d1-949b-467a-9f28-1dcc35ce3c8c',
-  },
-
-  m6I8TfROsyr5Wp73BcAd: {
-    _key: 'm6I8TfROsyr5Wp73BcAd',
-    _created: '2019-09-25T10:00:05.000Z',
-    _modified: '2019-09-25T11:00:01.000Z',
-    title: 'RAM Refresher',
-    mimetype: 'application/pdf',
-    description: '',
-    subtitle: '',
-    filename: 'ram-refresher.mp4',
-    type: 'file',
-    image: 'assets/resources/covers/ram-refresher.jpg',
-    url: 'https://firebasestorage.googleapis.com/v0/b/extension-toolkit.appspot.com/o/Resources%2Fram-refresher.mp4?alt=media&token=27939a71-0656-440b-8407-24877acaeede',
-    // youtubeID: 'Kw5UznKvCN8',
-  },
-};
-
-export default files;
