+++ apps/picsa-tools/resources-tool/src/app/data/picsa/manuals.ts
@@ -1,8 +1,8 @@
 import { IResourceCollection, IResourceFile } from '../../models';
 
 const manuals: Record<string, IResourceFile> = {
-  '4I4F36gD0MUef26HfPZG': {
-    _key: '4I4F36gD0MUef26HfPZG',
+  picsa_manual: {
+    _key: 'picsa_manual',
     _created: '2019-09-25T10:00:01.000Z',
     _modified: '2019-09-25T11:00:01.000Z',
     title: 'PICSA Manual (English)',
@@ -12,10 +12,10 @@ const manuals: Record<string, IResourceFile> = {
     subtitle: '',
     filename: 'picsa-field-manual-en.pdf',
     type: 'file',
-    url: 'https://firebasestorage.googleapis.com/v0/b/picsa-apps.appspot.com/o/picsa-field-manual-en.pdf?alt=media&token=6e2f1601-80c8-4f56-9a3d-52789f08c303',
+    url: 'https://firebasestorage.googleapis.com/v0/b/picsa-apps.appspot.com/o/picsa%2Fmanuals%2Fpicsa-manual.pdf?alt=media&token=8d5c21da-2252-44b7-8709-a3f12a593c2f',
   },
-  '9xqr6hxcWoVlsqydsNne': {
-    _key: '9xqr6hxcWoVlsqydsNne',
+  picsa_manual_chichewa: {
+    _key: 'picsa_manual_chichewa',
     _created: '2019-09-25T10:00:02.000Z',
     _modified: '2019-09-27T15:00:03.000Z',
     title: 'Buku la PICSA (Chichewa)',
@@ -25,11 +25,11 @@ const manuals: Record<string, IResourceFile> = {
     subtitle: 'Chichewa',
     filename: 'picsa-field-manual-chichewa.pdf',
     type: 'file',
-    url: 'https://firebasestorage.googleapis.com/v0/b/extension-toolkit.appspot.com/o/Manuals%2Fpicsa-field-manual-chichewa.pdf?alt=media&token=7ab629a9-d2c1-4b37-8237-88825033708e',
+    url: 'https://firebasestorage.googleapis.com/v0/b/picsa-apps.appspot.com/o/picsa%2Fmanuals%2Fpicsa-manual-chichewa.pdf?alt=media&token=d03031db-9e83-4a0a-a5be-7c3d1d42f98b',
     appCountries: ['mw', 'zm'],
   },
-  lqsSRwYx3Lq1bFYG3lke: {
-    _key: 'lqsSRwYx3Lq1bFYG3lke',
+  picsa_manual_tumbuka: {
+    _key: 'picsa_manual_tumbuka',
     _created: '2019-09-25T10:00:03.000Z',
     _modified: '2019-09-27T15:00:05.000Z',
     title: 'Buku la PICSA (Tumbuka)',
@@ -39,7 +39,7 @@ const manuals: Record<string, IResourceFile> = {
     subtitle: 'Tumbuka',
     filename: 'picsa-field-manual-tumbuka.pdf',
     type: 'file',
-    url: 'https://firebasestorage.googleapis.com/v0/b/extension-toolkit.appspot.com/o/Manuals%2Fpicsa-field-manual-tumbuka.pdf?alt=media&token=31a885e7-d6fa-4903-a747-ed8f9f7168b3',
+    url: 'https://firebasestorage.googleapis.com/v0/b/picsa-apps.appspot.com/o/picsa%2Fmanuals%2Fpicsa-manual-tumbuka.pdf?alt=media&token=ccaa5881-2200-48de-8799-69179b8109e0',
     appCountries: ['mw', 'zm'],
   },
   picsa_manual_tajik: {
@@ -53,12 +53,12 @@ const manuals: Record<string, IResourceFile> = {
     subtitle: '',
     filename: 'picsa_manual_tajik.pdf',
     type: 'file',
-    url: 'https://firebasestorage.googleapis.com/v0/b/picsa-apps.appspot.com/o/picsa-manuals%2Fpicsa_manual_tajik.pdf?alt=media&token=add3d42d-e6d5-46f0-b615-f5872228c210',
+    url: 'https://firebasestorage.googleapis.com/v0/b/picsa-apps.appspot.com/o/picsa%2Fmanuals%2Fpicsa_manual_tajik.pdf?alt=media&token=b8669e3c-9eae-4443-b743-6b1881990d81',
     appCountries: ['tj'],
   },
 };
 
-const collection: IResourceCollection = {
+const picsa_manuals: IResourceCollection = {
   _created: '2019-09-25T10:00:04.000Z',
   _modified: '2019-09-27T11:00:01.000Z',
   _key: 'picsaManual',
@@ -71,4 +71,4 @@ const collection: IResourceCollection = {
   childResources: Object.keys(manuals),
 };
 
-export default { ...manuals, collection };
+export default { ...manuals, picsa_manuals };
