+++ apps/picsa-tools/resources-tool/src/app/data/workshops/index.ts
@@ -1,24 +1,6 @@
 import { IResourceCollection } from '../../models';
 import mwSalima2022 from './mw-salima-2022';
-import zmChipata2022 from './zm-chipata-2022';
 
-const cropInfoSheets: IResourceCollection = {
-  _key: 'cropInfoSheets',
-  _created: '2019-09-25T10:00:04.000Z',
-  _modified: '2019-09-27T11:00:01.000Z',
-  title: 'Crop Information Sheets',
-  description: '',
-  subtitle: '',
-  type: 'collection',
-  image: 'assets/resources/covers/spreadsheet.svg',
-  parentResource: 'cropResources',
-  childResources: [
-    'crop_info_chipata',
-    'crop_info_petauke',
-    'crop_info_kasungu',
-    'crop_info_nkhotakota',
-  ],
-};
 const picsaWorkshops: IResourceCollection = {
   _created: '2022-09-20T10:00:04.000Z',
   _modified: '2022-09-20T11:00:01.000Z',
@@ -35,7 +17,5 @@ const picsaWorkshops: IResourceCollection = {
 
 export default {
   picsaWorkshops,
-  cropInfoSheets,
   ...mwSalima2022,
-  ...zmChipata2022,
 };
