+++ apps/picsa-tools/resources-tool/src/app/data/workshops/mw-salima-2022.ts
@@ -19,8 +19,6 @@ const collections: { [id: string]: IResourceCollection } = {
       'mw_seasonal_forecast_2022',
       'nkhotatkota_salima_advisory_31_july',
       'mw_weather_forecast_4th_september',
-      'crop_info_kasungu',
-      'crop_info_nkhotakota',
     ],
     parentResource: 'picsaWorkshops',
     appCountries: ['mw'],
@@ -95,34 +93,6 @@ const files: { [id: string]: IResourceFile } = {
     type: 'file',
     image: '',
     url: 'https://firebasestorage.googleapis.com/v0/b/picsa-apps.appspot.com/o/workshop-salima-2022%2FWEATHER%20FORECAST%20FOR%20TONIGHT%20AND%20TOMORROW_4TH%20SEPTEMBER%202022.pdf?alt=media&token=df20b361-0d14-4d69-82b4-8ca49d42d5aa',
-  },
-  crop_info_kasungu: {
-    _key: 'crop_info_kasungu',
-    _created: '2022-09-21T02:00:00.000Z',
-    _modified: '2022-09-21T02:00:00.000Z',
-    title: 'Crop Information Sheet',
-    mimetype: 'application/pdf',
-    description: 'Kasungu',
-    subtitle: '',
-    folder: 'workshop-salima-2022',
-    filename: 'CROP INFORMATION SHEET-KASUNGU.pdf',
-    type: 'file',
-    image: '',
-    url: 'https://firebasestorage.googleapis.com/v0/b/picsa-apps.appspot.com/o/workshop-salima-2022%2FCROP%20INFORMATION%20SHEET-KASUNGU.pdf?alt=media&token=2d70ebc5-62dd-41ab-bd5c-2977ae0b47ac',
-  },
-  crop_info_nkhotakota: {
-    _key: 'crop_info_nkhotakota',
-    _created: '2022-09-21T02:00:00.000Z',
-    _modified: '2022-09-21T02:00:00.000Z',
-    title: 'Crop Information Sheet',
-    mimetype: 'application/pdf',
-    description: 'Nkhotakota',
-    subtitle: '',
-    folder: 'workshop-salima-2022',
-    filename: 'NKHOTAKOTA Crop Information Sheet.pdf',
-    type: 'file',
-    image: '',
-    url: 'https://firebasestorage.googleapis.com/v0/b/picsa-apps.appspot.com/o/workshop-salima-2022%2FNKHOTAKOTA%20Crop%20Information%20Sheet.pdf?alt=media&token=a200f210-1ab5-45e5-9a94-753d2c225f8f',
   },
 };
