+++ apps/picsa-tools/resources-tool/src/assets/resources/contents.json
@@ -1,10 +1,4 @@
 {
-  "Agrometeorological Bulletin 1-10 April 2022.pdf": {
-    "relativePath": "Agrometeorological Bulletin 1-10 April 2022.pdf",
-    "size_kb": 419.6,
-    "md5Checksum": "9e719dd53e0f3b7f4f0143077dbabc8d",
-    "modifiedTime": "2022-09-21T12:19:51.027Z"
-  },
   "covers/access-ag.jpg": {
     "relativePath": "covers/access-ag.jpg",
     "size_kb": 34.8,
@@ -101,47 +95,29 @@
     "md5Checksum": "3fddfaa6aa5d40e2bd87edc834f9a63b",
     "modifiedTime": "2022-09-21T08:51:57.474Z"
   },
-  "CROP INFORMATION SHEET-KASUNGU.pdf": {
-    "relativePath": "CROP INFORMATION SHEET-KASUNGU.pdf",
-    "size_kb": 179.8,
-    "md5Checksum": "ae130f76d6498a96798323432125216f",
-    "modifiedTime": "2022-09-21T12:19:51.032Z"
-  },
   "crop-info-chipata.pdf": {
     "relativePath": "crop-info-chipata.pdf",
     "size_kb": 139.7,
     "md5Checksum": "89c26a2987e6e5688431c5c04eb17c25",
     "modifiedTime": "2022-09-30T09:15:38.622Z"
   },
+  "crop-info-kasungu.pdf": {
+    "relativePath": "crop-info-kasungu.pdf",
+    "size_kb": 179.8,
+    "md5Checksum": "ae130f76d6498a96798323432125216f",
+    "modifiedTime": "2022-10-13T23:03:57.360Z"
+  },
+  "crop-info-nkhotakota.pdf": {
+    "relativePath": "crop-info-nkhotakota.pdf",
+    "size_kb": 145.3,
+    "md5Checksum": "4bd29bdc95a4b4855550a3a1f17dd1b0",
+    "modifiedTime": "2022-10-13T23:07:40.308Z"
+  },
   "crop-info-petauke.pdf": {
     "relativePath": "crop-info-petauke.pdf",
     "size_kb": 140.1,
     "md5Checksum": "b58f4cdf1c67f97c999b32abec8e3ee3",
     "modifiedTime": "2022-09-30T09:15:14.200Z"
-  },
-  "crop-info-sheet-chileka-cover.png": {
-    "relativePath": "crop-info-sheet-chileka-cover.png",
-    "size_kb": 10.1,
-    "md5Checksum": "3a6935b6986b73243a641e1b82376190",
-    "modifiedTime": "2022-09-08T10:31:19.010Z"
-  },
-  "Malawi Seasonal forecast 2021-2022 Rainfall season.pdf": {
-    "relativePath": "Malawi Seasonal forecast 2021-2022 Rainfall season.pdf",
-    "size_kb": 567.8,
-    "md5Checksum": "2f8d061804323d651297acfa773582f4",
-    "modifiedTime": "2022-09-21T12:19:51.041Z"
-  },
-  "Nkhotakota & Salima Issued on 31 July 2022.docx.pdf": {
-    "relativePath": "Nkhotakota & Salima Issued on 31 July 2022.docx.pdf",
-    "size_kb": 173.6,
-    "md5Checksum": "ca985de1a62c9ee6cebfdccdd521032e",
-    "modifiedTime": "2022-09-21T12:19:51.054Z"
-  },
-  "NKHOTAKOTA Crop Information Sheet.pdf": {
-    "relativePath": "NKHOTAKOTA Crop Information Sheet.pdf",
-    "size_kb": 145.3,
-    "md5Checksum": "4bd29bdc95a4b4855550a3a1f17dd1b0",
-    "modifiedTime": "2022-09-21T12:19:51.045Z"
   },
   "picsa-field-manual-en.pdf": {
     "relativePath": "picsa-field-manual-en.pdf",
@@ -160,17 +136,5 @@
     "size_kb": 17.8,
     "md5Checksum": "1431a9389a8192ac6bbf52702e199fb0",
     "modifiedTime": "2022-09-08T10:31:19.250Z"
-  },
-  "seasonal-forecast-mangochi-2017-2018.jpg": {
-    "relativePath": "seasonal-forecast-mangochi-2017-2018.jpg",
-    "size_kb": 113.1,
-    "md5Checksum": "3f53b3436b38b9d14779c3d9b9a6e1bd",
-    "modifiedTime": "2022-09-08T10:31:19.253Z"
-  },
-  "WEATHER FORECAST FOR TONIGHT AND TOMORROW_4TH SEPTEMBER 2022.pdf": {
-    "relativePath": "WEATHER FORECAST FOR TONIGHT AND TOMORROW_4TH SEPTEMBER 2022.pdf",
-    "size_kb": 157.9,
-    "md5Checksum": "e2ebf8fe86405570fb978acc74e08ecb",
-    "modifiedTime": "2022-09-21T12:19:51.059Z"
   }
 }
