+++ apps/picsa-tools/resources-tool/src/assets/resources/crop-info-kasungu.pdf
