+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "picsa-apps",
-  "version": "3.9.0",
+  "version": "3.10.0",
   "license": "See LICENSE",
   "scripts": {
     "ng": "nx",
